started jmonkey ui

+++ trunk/src/dev/vbms/autopilot/sim/device/AbstractThermopile.java
@@ -1,17 +1,9 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
 package dev.vbms.autopilot.sim.device;
 
-import dev.vbms.autopilot.device.*;
+import com.jme3.math.Vector3f;
 
-/**
- *
- * @author pcmobile
- */
 public abstract class AbstractThermopile extends AbstractDevice {
     public abstract int[][] getData ();
-    public abstract Vector3d getPosition ();
-    public abstract Vector3d getDirection ();
+    public abstract Vector3f getPosition ();
+    public abstract Vector3f getDirection ();
 }
